# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MAUI-API
  version: 1.0.0-draft
  summary: Backend for MAUI arcade cabinets running in ONLINE mode.
  description: |
    Lot 1 contract: MAUI client authentication, machine binding and telemetry.
    Lot 2 (in progress): game catalog, players.
    See `docs/PLAN.md` for the design decisions behind this contract.

    ## Authentication

    Every `/api/v1` request from a MAUI cabinet sends three headers:

    | Header | Content |
    |--------|---------|
    | `X-Maui-Key` | Public client key (`mk_...`), stable across renewals. |
    | `Authorization` | `Bearer <token>`, the secret. Only one valid token per client. |
    | `X-Maui-Machine` | Machine fingerprint, see below. |

    All three are required together. A token that does not belong to the client
    carrying the key is rejected as if it were unknown.

    ## Machine fingerprint

    MAUI computes the fingerprint locally and never receives it from the API:

    `fingerprint = lowercase hex SHA-256 of (local UUID + ":" + OS machine identifier)`

    - local UUID: generated once, persisted in the MAUI config directory;
    - OS machine identifier: `/etc/machine-id` (Linux), `IOPlatformUUID` (macOS),
      `MachineGuid` (Windows). Empty string if unavailable.

    The fingerprint must be computed by the Electron main process or the BO
    Express server, never by a browser page (the BO is reachable from the LAN).

    The first authenticated request on an unbound client binds the key to the
...

...
servers:
...
security:
  - mauiKey: []
    bearerToken: []
    mauiMachine: []
paths:
...
  /players/link:
    post:
      tags: [Players]
      operationId: linkPlayer
      summary: Link an existing player with its PIN
      description: |
        Links the player owning these initials to this cabinet. Idempotent for
        a player already linked (the PIN is still checked). Each wrong PIN is
        counted; the fifth in a row locks the player (`423`) until a cabinet
        of this player issues a new PIN, or an admin unlocks it. Own rate
        limit: 10 requests per minute per key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkPlayerRequest'
      responses:
        '200':
          description: Player linked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerResponse'
        '404':
          $ref: '#/components/responses/PlayerNotFound'
        '423':
          $ref: '#/components/responses/PlayerLocked'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '400':
          $ref: '#/components/responses/MachineFingerprintMissing'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PlayerForbidden'
        '409':
          $ref: '#/components/responses/MachineMismatch'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /players/{playerId}/link:
    parameters:
      - $ref: '#/components/parameters/PlayerId'
    delete:
      tags: [Players]
      operationId: unlinkPlayer
      summary: Unlink a player from this cabinet
      description: |
        The player keeps its initials and its other cabinets.
      responses:
        '204':
          description: Unlinked.
        '404':
          $ref: '#/components/responses/PlayerNotFound'
        '400':
          $ref: '#/components/responses/MachineFingerprintMissing'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PlayerForbidden'
        '409':
          $ref: '#/components/responses/MachineMismatch'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    mauiKey:
      type: apiKey
      in: header
      name: X-Maui-Key
      description: Public client key, format `mk_` followed by 24 base62 characters.
    bearerToken:
      type: http
      scheme: bearer
      description: Sanctum token, opaque string. Never displayed twice.
    mauiMachine:
      type: apiKey
      in: header
      name: X-Maui-Machine
      description: Machine fingerprint, 64 lowercase hex characters (SHA-256).
    frontSession:
      type: apiKey
      in: cookie
      name: maui-front-session
      description: |
        Session of a member of the hiscores front, set by the Discord login
        (D64). A cookie of its own, apart from the back office's.
  parameters:
    PlayerId:
      name: playerId
      in: path
      required: true
      description: Player `id` as returned by the API (UUID).
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Pseudo3:
      type: string
      pattern: '^[A-Z]{3}$'
      description: Player initials, 3 uppercase letters (same rule as MAUI's Pseudo3.ts).
      examples: [ACE]
    Player:
      type: object
      required: [id, pseudo_3, is_public, status, is_origin, avatar]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        pseudo_3:
          $ref: '#/components/schemas/Pseudo3'
        is_public:
          type: boolean
        status:
          type: string
          enum: [active, disabled, locked]
          description: '`disabled` by an admin; `locked` after too many wrong PINs.'
        is_origin:
          type: boolean
          description: The player was created on this cabinet, the only one that can issue a new PIN (D54).
        avatar:
          type: [string, 'null']
          description: SHA-256 of the player's avatar (D53), null without one. Send the PNG again when the cabinet's differs.
    PlayerResponse:
      type: object
      required: [player]
      additionalProperties: false
      properties:
        player:
          $ref: '#/components/schemas/Player'
...
    Pin:
      type: string
      pattern: '^[0-9]{4}$'
      description: 4-digit PIN, returned once in plain text.
...
    LinkPlayerRequest:
      type: object
      required: [pseudo_3, pin]
      additionalProperties: false
      properties:
        pseudo_3:
          $ref: '#/components/schemas/Pseudo3'
        pin:
          $ref: '#/components/schemas/Pin'
...
    Problem:
      type: object
      description: RFC 9457 problem details, with a stable `code` extension.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        attempts_left:
          type: integer
          description: '`pin_invalid` only: wrong PINs left before the player is locked.'
        code:
          type: string
          enum:
            - machine_fingerprint_missing
            - unauthenticated
            - client_disabled
            - insufficient_ability
            - machine_mismatch
            - validation_failed
            - rate_limited
            - server_error
            - initials_taken
            - player_not_found
            - player_disabled
            - player_locked
            - pin_invalid
            - not_origin_cabinet
            - avatar_not_found
            - game_not_found
            - media_not_found
            - origin_not_allowed
            - member_disabled
            - invitation_unavailable
            - player_already_linked
            - member_has_player
            - forbidden
            - not_found
            - method_not_allowed
            - http_error
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required: [errors]
          properties:
            errors:
              type: object
              description: Field name to list of messages.
              additionalProperties:
                type: array
                items:
                  type: string
  responses:
    MachineFingerprintMissing:
      description: '`X-Maui-Machine` header missing or not 64 lowercase hex characters.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Bad Request
            status: 400
            code: machine_fingerprint_missing
    Unauthenticated:
      description: Key or token missing, unknown, revoked, or not matching each other.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Unauthorized
            status: 401
            code: unauthenticated
...
    MachineMismatch:
      description: Client already bound to another machine fingerprint.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Conflict
            status: 409
            detail: These credentials are already used on another cabinet.
            code: machine_mismatch
    ValidationFailed:
      description: Request body invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
          example:
            type: about:blank
            title: Unprocessable Content
            status: 422
            code: validation_failed
            errors:
              os:
                - The selected os is invalid.
    RateLimited:
      description: Too many requests for this client key.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Too Many Requests
            status: 429
            code: rate_limited
    PlayerForbidden:
      description: |
        Client disabled, token lacking the `players` ability, player disabled
        by an admin, wrong PIN (`attempts_left` before the lock), or a new
        PIN asked from a cabinet that is not the player's origin.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            clientDisabled:
              value:
                type: about:blank
                title: Forbidden
                status: 403
                code: client_disabled
            playerDisabled:
              value:
                type: about:blank
                title: Forbidden
                status: 403
                code: player_disabled
            pinInvalid:
              value:
                type: about:blank
                title: Forbidden
                status: 403
                code: pin_invalid
                attempts_left: 3
            notOriginCabinet:
              value:
                type: about:blank
                title: Forbidden
                status: 403
                code: not_origin_cabinet
...
    PlayerNotFound:
      description: No player with these initials, or not linked to this cabinet.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Not Found
            status: 404
            code: player_not_found
    PlayerLocked:
      description: Too many wrong PINs in a row.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Locked
            status: 423
            code: player_locked
    ServerError:
      description: Unexpected error. No technical detail exposed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Internal Server Error
            status: 500
            code: server_error